requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-anymail==7.1.0
django-ckeditor==5.9.0
django-environ==0.4.5
django-hitcount==1.3.2
django-import-export==2.1.0
django-js-asset==1.2.2
django-recaptcha2==1.4.1
django-taggit==1.3.0
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
